+++ packages/react-router/src/components/Redirect.tsx
@@ -1,17 +1,21 @@
 import React, {useEffect} from 'react';
-import {withRouter, RouteComponentProps} from 'react-router';
+import {withRouter, RouteComponentProps} from 'react-router-dom';
 import {Redirect as NetworkRedirect} from '@shopify/react-network';
 
 interface Props {
   url: string;
 }
 
-type ComposedProps = Props & RouteComponentProps<{}, {}>;
+type ComposedProps = Props & RouteComponentProps;
 
-function Redirect(props: ComposedProps) {
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  useEffect(() => props.router.push(props.url), []);
-  return <NetworkRedirect url={props.url} />;
+function Redirect({url, history}: ComposedProps) {
+  useEffect(
+    () => {
+      history.push(url);
+    },
+    [history, url],
+  );
+  return <NetworkRedirect url={url} />;
 }
 
-export default withRouter<Props>(Redirect);
+export default withRouter(Redirect);