+++ packages/react-router/CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Fix Redirect component
+
 ## [0.0.4] - 2019-09-05
 
 - Move the types to depenedencies